+++ internal/handlers/message.go
@@ -1,8 +1,8 @@
 package handlers
 
 import (
+	"github.com/Ayano2000/push/internal/pkg/logger"
 	"github.com/Ayano2000/push/internal/pkg/types"
-	"github.com/Ayano2000/push/pkg/logger"
 	"github.com/Ayano2000/push/pkg/transformer"
 	"io"
 	"net/http"
