+++ internal/handlers/webhooks.go
@@ -2,8 +2,8 @@ package handlers
 
 import (
 	"encoding/json"
+	"github.com/Ayano2000/push/internal/pkg/logger"
 	"github.com/Ayano2000/push/internal/pkg/types"
-	"github.com/Ayano2000/push/pkg/logger"
 	"github.com/Ayano2000/push/pkg/transformer"
 	"github.com/pkg/errors"
 	"net/http"
